Provide a job timeout for the package stage
Because the package publish is a job it was provided the default of 60 minutes.
However the test/package stages can exceed that period - meaning the
package job would exceed the default timeout and fail before it had an
opportunity to execute.

This change parameterizes the timeout and adds it to the package job.